# docs/swagger/swagger.yaml
basePath: /
definitions:
  go-micro_internal_activity.CouponRequest:
    properties:
      amount:
        type: number
      coupon_id:
        type: string
      user_id:
        type: string
    type: object
  go-micro_internal_activity.SeckillRequest:
    properties:
      quantity:
        type: integer
      sku_id:
        type: string
      user_id:
        type: string
    type: object
...
  go-micro_pkg_httpx.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Go-Micro 网关服务 API
  title: Go-Micro Gateway API
  version: "1.0"
paths:
  /api/v1/activity/coupon:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: issue coupon
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-micro_internal_activity.CouponRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: Issue coupon
      tags:
      - Activity
  /api/v1/activity/seckill:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: seckill request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-micro_internal_activity.SeckillRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: Seckill request
      tags:
      - Activity
  /api/v1/activity/status:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: coupon id
        in: query
        name: coupon_id
        type: string
      - description: sku id
        in: query
        name: sku_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: Activity status
      tags:
      - Activity
...
swagger: "2.0"